+++ xugrid/__init__.py
@@ -39,7 +39,7 @@ from xugrid.ugrid.snapping import (
 from xugrid.ugrid.ugrid1d import Ugrid1d
 from xugrid.ugrid.ugrid2d import Ugrid2d
 
-__version__ = "0.15.1"
+__version__ = "0.15.2"
 
 __all__ = (
     "data",